Generate a QSLIMFinder command

Generate a QSLIMFinder command

QSLIMFinderCommand(file_list, i = 1,
  slimpath = "../software/slimsuite/tools/",
  blast = "../software/ncbi_blast_2.6.0/bin/",
  iupred = "../software/iupred/iupred",
  options = "dismask=T consmask=F cloudfix=F probcut=0.1",
  LSF_cluster_par = "bsub -n 1 -q research-rh7 -M 100 -R \"rusage[mem=100]\"",
  LSF_project_path = "/hps/nobackup/research/petsalaki/users/vitalii/vitalii/viral_project/",
  analysis_type = c("qslimfinder", "slimfinder")[1])

mQSLIMFinderCommand(file_list, slimpath = "../software/slimsuite/tools/",
  blast = "../software/ncbi_blast_2.6.0/bin/",
  iupred = "../software/iupred/iupred",
  options = "dismask=T consmask=F cloudfix=F probcut=0.5",
  LSF_cluster_mode = F,
  LSF_project_path = "/hps/nobackup/research/petsalaki/users/vitalii/vitalii/viral_project/",
  LSF_cluster_par = "bsub -n 1 -q research-rh7 -M 200 -R \"rusage[mem=200]\"",
  log_dir = "./SLIMFinder/log_dir/", write_log = T, recursive = F,
  analysis_type = c("qslimfinder", "slimfinder")[1])

groupQSLIMFinderCommand(commands, InteractionSubsetFASTA_list,
  sh_dir = "/sh_dir/",
  LSF_project_path = "/hps/nobackup/research/petsalaki/users/vitalii/vitalii/viral_project/",
  dataset_name = "SLIMFinder", N_seq = 200, write_log = T)

Arguments

file_list
data.table containing path to files and directories for QSLIMFinder: fastafile, queryfile, outputdir, outputfile

i
integer, which set of files and directories to choose from file_list

slimpath
relative path (from the project folder) to the directory containing qslimfinder.py or slimfinder.py

blast
relative path (from the project folder) to /bin/ folder in BLAST package

iupred
relative path (from the project folder) to iupred (compiled executable)

options
any options from QSLIMFinder

LSF_cluster_par
a string that will launch LSF job

LSF_project_path
absolute path on LSF cluster

analysis_type
character 1L, qslimfinder or slimfinder. slimfinder doesn't need query: identical datasets with the same query will be removed and query files not written)

LSF_cluster_mode
logical, if FALSE LSF_cluster_par and LSF_project_path are ignored

log_dir
character, directory where to write log files (stout and sterr)

write_log
whether to save stout and sterr

commands
list returned by mQSLIMFinderCommand() containing: 1. command to set up enviromental variable IUPred_PATH; 2. character vector of bash commands that will lauch QSLIMFinder as a job on LSF cluster; 3, 4, 5 - directories where LSF should write stout and sterr

InteractionSubsetFASTA_list
object of class InteractionSubsetFASTA_list containing: FASTA sequences for interacting proteins, molecular interaction data they correspond to. Each element of a list contains input for individual QSLIMFinder run.

sh_dir
directory within dataset directory (dataset_name) where to write batch command .sh files

dataset_name
character, name of the dataset, such as "SLIMFinder" or "SLIMFinder_Vidal"

N_seq
size of the batch. Groups QSLIMFinder jobs until the next job doesn't fit into N_seq, if jobs is larger than N_seq a single job will be written to a batch .sh file.

LSF_project_path
full path to the project where dataset directory (dataset_name) is located.

write_log
FALSE will not allow runQSLIMFinder to detect crashed jobs

Value

character vector (1L), bash command that will lauch QSLIMFinder locally or as a job on LSF cluster

list containing: 1. command to set up enviromental variable IUPred_PATH; 2. character vector of bash commands that will lauch QSLIMFinder as a job on LSF cluster; 3, 4, 5 - directories where LSF should write stout and sterr

QSLIMFinderCommand split into batches by N_seq sequences. List containing: 1. command to set up enviromental variable IUPred_PATH; 2. character vector of bash commands that will lauch $SHELL as a job on LSF cluster and run QSLIMFinder commands from batch file; 3, 4, 5 - directories where LSF should write stout and sterr

See also

Examples

all_commands = mQSLIMFinderCommand(file_list = forSLIMFinder_file_list,
    slimpath = "../software/cluster/slimsuite/tools/qslimfinder.py",
    blast = "../software/cluster/ncbi_blast_2.6.0/bin/",
    iupred = "../software/cluster/iupred/iupred",
    options = "dismask=T consmask=F cloudfix=T probcut=0.3 iuchdir=T",
    LSF_cluster_mode = T,
    LSF_project_path = "/hps/nobackup/research/petsalaki/users/vitalii/vitalii/viral_project/",
    LSF_cluster_par = "bsub -n 1 -q research-rh7 -M 100 -R \"rusage[mem=100]\"")

#> Error in nrow(file_list): object 'forSLIMFinder_file_list' not found