Plot ROC or precision-recall curves for motifs search benchmark

Plot ROC or precision-recall curves for motifs search benchmark

benchmarkMotifsROC(res, data_type = c("both", "query", "all")[1],
  col_query = c("grey"), col_all = c("black"), ...,
  text_args = "col = \"black\"", ROCR_data = F,
  legend_args = "x = 0.8 | y = 0.25", x_query = 0.25, x_all = 0.75,
  measure1 = "tpr", measure2 = "fpr", single_metric = c("auc",
  "prbe")[1], single_metric_name = c("Median AUC",
  "Precision-recall break-even point")[1],
  summarise_single_metric = c("median", "")[1],
  both_metrics_vs_cutoff = F)

mBenchmarkMotifsROC(res_list, data_type = "query", ...)

Arguments

res
object class benchmarkMotifsResult, the output of benchmarkMotifs

...
args passed to plot

text_args
character 1L, arguments to text and legend that allow to manipulate the median AUC and legend on the plot

ROCR_data
logical, return ROC performance data? If FALSE the function only plots ROC curve

legend_args
character 1L, arguments to text and legend that allow to manipulate the median AUC and legend on the plot

measure1
character, for example "tpr" or "prec", details: ?ROCR::performance

measure2
character, for example "fpr" or "rec", details: ?ROCR::performance

res_list
list of objects of class benchmarkMotifsResult, the output of mBenchmarkMotifs

type
benchmark (test if known motifs were observed) in query proteins or in all proteins the sequences of which were used in QSLIMFinder search

Value

benchmarkMotifsROC: list of ROCR objects (prediction, performance, auc.performance) used to plot ROC curve

mBenchmarkMotifsROC: list of lists of ROCR objects (prediction, performance, auc.performance) used to plot ROC curve

See also