Center QSLIMFinder datasets (InteractionSubsetFASTA_list) at domains

Center QSLIMFinder datasets (InteractionSubsetFASTA_list) at domains

centerDomains(interactionFASTA_list, domain_res,
  domain_col = "IDs_domain_human", query_col = "IDs_interactor_viral",
  seed_col = "IDs_interactor_human")

Arguments

interactionFASTA_list
object of class InteractionSubsetFASTA_list containing: FASTA sequences for interacting proteins, molecular interaction data they correspond to. Each element of a list contains input for individual QSLIMFinder run.

domain_res
domain enrichment results (object of S3 class "XYZinteration_XZEmpiricalPval" (list), the output of permutationPval() call), or data.table with domain_col, query_col and seed_col

domain_col
name of the column in domain_res containing domain ID, not relevant if domain_res is XYZinteration_XZEmpiricalPval

query_col
name of the column in domain_res containing query protein ID, not relevant if domain_res is XYZinteration_XZEmpiricalPval

seed_col
name of the column in domain_res containing seed protein ID, not relevant if domain_res is XYZinteration_XZEmpiricalPval

Value

object of class InteractionSubsetFASTA_list containing domain-centered datasets

See also

Examples

forSLIMFinder = listInteractionSubsetFASTA(interaction_set1 = all_human_interaction,
                 interaction_set2 = all_viral_interaction,
                 seed_id_vect = proteins_w_signif_domains,
                 fasta = all.fasta,
                 single_interact_from_set2 = T, set1_only = T)

#> Error in grepl("clean_MItab", class(interaction_set1)): object 'all_human_interaction' not found