Process molecular interaction data in MITAB 2.5 or 2.7 formats

cleanMITAB extracts columns of molecular interaction data in MITAB 2.5 or 2.7 formats. For simple MITAB 2.5, columns are interactor Uniprot IDs(or other ID as specified by interactor_IDs_databases), interactor taxonomy IDs, Publication Identifiers, Confidence values and generates unique pair ID. MITAB 2.7 includes additional information: see details.

cleanMITAB(MITABdata)

Arguments

MITABdata
object of class "RAW_MItab25" or "RAW_MItab27" (list) containing molecular interaction data as returned by queryPSICQUICrlib

Value

Object of S3 class "clean_MItab25" or "clean_MItab27" containing interaction data in a data.table. Within each interacting pair all interactor attributes are sorted according to alphanumerically sorted interactor UniprotKB accessions

Details

Output column description (MITAB 2.5):

pair_id - unique identifier of the undirected interaction: ordered alphabetically and concatenated interacting molecule IDs

IDs_interactor_A, IDs_interactor_B - interacting molecule ID

interactor_IDs_databases_A, interactor_IDs_databases_B - database that provides interacting molecule ID such as UniProt, ChEMBL or IntAct (IntAct: when the molecule cannot be mapped to ID in any other resource)

Taxid_interactor_A, Taxid_interactor_B - taxonomic species that interacting molecule belongs to

Publication_Identifiers - pubmed ID of the papers that has reported the interaction

Confidence_values - MIscore. Details: https://psicquic.github.io/MITAB25Format.html, http://www.ebi.ac.uk/intact/pages/faq/faq.xhtml, https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4316181/

if MITAB 2.7 then the following columns are added:

Host_organisms - taxid of organism in which the interaction detections experiment was performed

bait_prey_status_A, bait_prey_status_B - specifies the experimental role of a protein: a bait, a prey, a neutral component or else (http://www.ebi.ac.uk/ols/ontologies/MI/terms?obo_id=MI:0495)

Interaction_detection_methods - Method to determine the interaction. http://www.ebi.ac.uk/ols/ontologies/MI/terms?obo_id=MI:0001

Interaction_types - http://www.ebi.ac.uk/ols/ontologies/MI/terms?obo_id=MI:0190

Interaction_identifiers - Interaction identifier given by database (e.g. intact:EBI-1547321|imex:IM-11865-3)

Expansion_methods - The method by which complex n-ary data is expanded into binary data. http://www.ebi.ac.uk/ols/ontologies/MI/terms?obo_id=MI:1059

Features_interactor_A, Features_interactor_B - Property of a subsequence that may interfere with the binding of a molecule. http://www.ebi.ac.uk/ols/ontologies/MI/terms?obo_id=MI:0116

Identification_method_participant_A, Identification_method_participant_B - Method to determine the molecules involved in the interaction. http://www.ebi.ac.uk/ols/ontologies/MI/terms?obo_id=MI:0002

binding_region_A_start, binding_region_A_end, binding_region_B_start, binding_region_B_end - start and end position of binding-associated features

binding_region_A_type, binding_region_B_type - Type of the binding region. Details: http://www.ebi.ac.uk/ols/ontologies/MI/terms?obo_id=MI:0117, https://psicquic.github.io/MITAB27Format.html, MITABregionFeature

See also