Find the degree distribution given edgelist (pair_id) and a list of proteins which degrees matter

Find the degree distribution given edgelist (pair_id) and a list of proteins which degrees matter

edgelist2degreeFreq(mitab, prots = NULL)

Arguments

mitab
data.table containing molecular interaction data, including pair_id

prots
list of relevant proteins, degree distribution is based on the prevalence of specific degrees in this list, all proteins in mitab by default

sep
ID separator in pair_id, passed to edgelist2degree

Details

unique pair_id is the unique identifier of the undirected interaction: ordered alphabetically and concatenated interacting molecule IDs cleanMITAB

Examples

if(F) {
# download full human interactome (clean = TRUE is necessary to produce the right input for edgelist2degree)
full = fullInteractome(taxid = "9606", database = "IntActFTP", format = "tab25", clean = TRUE, protein_only = TRUE)
degree_freq = edgelist2degreeFreq(full)
plot(degree_freq$N, degree_freq$degree_freq,
     log = "xy",
     main = "degree distribution \n - human-human protein interactions -",
     ylab = "degree frequency, log-scale",
     xlab = "degree, log-scale",
     las=1)
}