Find occuring GRanges overlapping with the benchmarking set (positive + negative examples)

Find occuring GRanges overlapping with the benchmarking set (positive + negative examples)

findOverlapsBench(occuring, benchmarking, predictor_col = "Sig",
  labels_col = "for_benchmarking", normalise = T,
  select_predictor_per_range = min, ...)

Arguments

occuring
GRanges object, occuring feature instances

benchmarking
GRanges object, feature instances for benchmarking

predictor_col
name of the metadata column in occuring that contains 1 - predictor value (such as p-value)

labels_col
name of the metadata column in benchmarking that contains labels for benchmarking (for, example 0 for negative and 1 for positive)

normalise
logical, normalise predictor value, just in case predictor doesn't span the full range between 0 ... 1

...
arguments such as maxgap, minoverlap passed to findOverlaps

Value

Genomic Ranges list object containing N randomised GRanges objects

See also