find fold_enrichment distribution (in any domain or other feature) for each protein in the network

foldEnrichmentDist finds fold_enrichment distribution (in any domain or other feature) for each protein in the first column of net. For more details: foldEnrichment

{
library(data.table)
data = fread("https://raw.githubusercontent.com/vitkl/viral_project/master/processed_data_files/viral_human_net_w_domains")
# generate minimal information tables
net = unique(data[,.(IDs_interactor_viral, IDs_interactor_human, IDs_interactor_viral_degree)])
protein_annot = unique(data[,.(IDs_interactor_human, IDs_domain_human, domain_frequency)])
viral_human_net_w_domains = unique(data[,.(IDs_interactor_viral, IDs_interactor_human, IDs_domain_human, domain_frequency_per_IDs_interactor_viral, fold_enrichment)])
foldEnrichmentDist(net, protein_annot, N = 10, cores = NULL, seed = 1, frequency = T)
}

Arguments

net
data.table that specifies the PPI network: 1. first column specifies a protein ("IDs_interactor_viral") 2. second column specifies it's interacting partner which annotations are of interest ("IDs_interactor_human") 3. third column specifies the degree of a protein in the first column ("IDs_interactor_viral_degree")

protein_annot
data.table that specifies annotations for a protein in the second column of net 1. first column specifies a protein ("IDs_interactor_human") 2. second column specifies features of that protein ("IDs_domain_human") 3. third column specifies background frequency of those features ("domain_frequency)

N
number of times to run permutation of PPI network

cores
specify how many cores to use for parallel processing, default (NULL) is to detect all cores on the machine and use all minus one. When using LSF cluster you must specify the number of cores to use because detectCores doen't know how much cores you have requested from LSF (with bsub -n) and detects all cores on the actual physical node.

seed
seed for RNG for reproducible sampling

frequency
fold enrichment or frequency in a set (if TRUE - frequency), also passed to foldEnrichment

Value

data.table containing fold enrichment for each domain (or other feature) - protein pair (2 columns: IDs_interactor_viral and sampled_fold_enrichment or sampled_domain_frequency_per_set)