Find pvalue for each protein-feature pair in the network

foldEnrichmentPval calculates how often each fold Enrichment value for the protein-feature pair is as high or higher than the foldEnrichment value for the protein - any (random) feature pairs (obtained from permutations using foldEnrichmentDist).

plotFoldEnrichmentDist allows to visualise sampled fold Enrichment distribution and the real fold Enrichment for the protein-feature pair for specific protein

foldEnrichmentPval(fold_enrichment_dist, data, cores = NULL, frequency = T)
plotFoldEnrichmentDist(proteinID, fold_enrichment_dist, data, main = NULL, text_lab = T, frequency = T)

Arguments

fold_enrichment_dist
data.table containing fold enrichment for each domain (or other feature) - protein pair (2 columns: IDs_interactor_viral and sampled_fold_enrichment), generated by foldEnrichmentDist

data
data.table containing foldEnrichment value for each protein-feature pair (4 columns: IDs_interactor_viral, IDs_interactor_human, IDs_domain_human, fold_enrichment)

cores
specify how many cores to use for parallel processing, default (NULL) is to detect all cores on the machine and use all minus one. When using LSF cluster you must specify the number of cores to use because detectCores doen't know how much cores you have requested from LSF (with bsub -n) and detects all cores on the actual physical node.

frequency
fold enrichment or frequency in a set (if TRUE - frequency)

proteinID
character specifying proteinID for which to plot fold Enrichment distribution

main
title for the plot

text_lab
logical: label domains (features) on the plot, or not

Value

data.table containing pvalue for each protein-feature pair in the network (4 columns: proteinID, IDs_interactor_human, featureID, foldEnrichment (or domain_frequency_per_IDs_interactor_viral) and Pval)