Compare viral and human protein degree distribution, filter by interaction detection method or PMIDs

subset molecular interaction data (cleaned MITAB format in a data.table object) with a list of interactors

humanViralDegree(data = NULL, directory = "./data_files/",
  Interaction_detection_methods = NULL, Identification_method = NULL,
  PMIDs = NULL, inverse_filter = F, data_name = "")

Arguments

data
a list contatining PPI data as generated by loadHumanViralPPI, if NULL, data will be downloaded internally

directory
where to find / keep PPI data (arg for fullInteractome, interSpeciesInteractome)) and MI ontology (arg for subsetMITABbyMethod)

Interaction_detection_methods
arg for subsetMITABbyMethod

Identification_method
arg for subsetMITABbyMethod

PMIDs
arg for subsetMITABbyPMIDs

inverse_filter
logical, inverse filtering criteria

data_name
to be displayed on plot

Value

data.table to be ggplotted to compare viral and human protein degree distributions

Examples

# Load data. This can be done internally by humanViralDegree(),
# however, it takes time (minutes) so it's recommended to load data separately
# if it will be used multiple times
HumanViralPPI = loadHumanViralPPI(directory = "./data_files/")

#> Error in loadIntActFTP(dir = loadIntActFTP_dir, release = release): no data for IntAct release2017Nov13 in the directory: ./data_files/IntActRelease_2017Nov13/, set release to NULL do load the latest release

# prepare all data for the degree distribution analysis
degree_distributions = humanViralDegree(data = HumanViralPPI,
  directory = "./data_files/", data_name = "Full IntAct")

#> Error in humanViralDegree(data = HumanViralPPI, directory = "./data_files/",     data_name = "Full IntAct"): object 'HumanViralPPI' not found

# prepare two-hybrid data for the degree distribution analysis
degree_distributions = humanViralDegree(data = HumanViralPPI,
  directory = "./data_files/", Interaction_detection_methods = "MI:0018",
  data_name = "two-hybrid data")

#> Error in humanViralDegree(data = HumanViralPPI, directory = "./data_files/",     Interaction_detection_methods = "MI:0018", data_name = "two-hybrid data"): object 'HumanViralPPI' not found

# prepare AP-MS data for the degree distribution analysis
degree_distributions = humanViralDegree(data = HumanViralPPI,
  directory = "./data_files/", Interaction_detection_methods = "MI:0004",
  Identification_method = "MI:0433", PMIDs = NULL, inverse_filter = F,
  data_name = "AP-MS data")

#> Error in humanViralDegree(data = HumanViralPPI, directory = "./data_files/",     Interaction_detection_methods = "MI:0004", Identification_method = "MI:0433",     PMIDs = NULL, inverse_filter = F, data_name = "AP-MS data"): object 'HumanViralPPI' not found

# to see if viral-interacting human proteins are special in the human network,
# look at the Vidal published and unpublished datasets (only human-human network
# is modified (data for the top 2 plots))
degree_distributions = humanViralDegree(data = HumanViralPPI,
  directory = "./data_files/", PMIDs = c("25416956", "unassigned1304"),
  data_name = "Vidal published and unpublished")

#> Error in humanViralDegree(data = HumanViralPPI, directory = "./data_files/",     PMIDs = c("25416956", "unassigned1304"), data_name = "Vidal published and unpublished"): object 'HumanViralPPI' not found

# to see if viral-interacting human proteins are special in the human network,
# do the same for Matthias Mann 2015 paper dataset
degree_distributions = humanViralDegree(data = HumanViralPPI,
  directory = "./data_files/", PMIDs = "26496610",
  data_name = "Matthias Mann 2015 paper")

#> Error in humanViralDegree(data = HumanViralPPI, directory = "./data_files/",     PMIDs = "26496610", data_name = "Matthias Mann 2015 paper"): object 'HumanViralPPI' not found