Retrieve interspecies interactions (proteins only or not) for a given pair of taxonomy IDs

Retrieve interspecies interactions (proteins only or not) for a given pair of taxonomy IDs from a specified database. Interactome can be additionally cleaned and includes only specific information: cleanMITAB

interSpeciesInteractome(MITABdata = NULL, taxid1 = 9606,
  taxid2 = 10239, database = "imex", format = "tab25",
  clean = TRUE, protein_only = TRUE, directory = NULL,
  releaseORdate = NULL, remove_obsolete_id = F)

Arguments

MITABdata
object of class "RAW_MItab25" or "RAW_MItab27" (list) containing molecular interaction data as returned by queryPSICQUICrlib or loadIntActFTP, default in NULL

taxid1
character (1L), taxonomy id of the species which interaction participants should belong to, default is 9606 (which is human)

taxid2
character (1L), taxonomy id of the species which interaction participants should belong to, default is 10239 (which is all viral taxa)

database
character (1L), argument for queryPSICQUIC, PSICQUIC-compliant database to query for interactions. The default is "imex" alternative to which is "IntActFTP"

format
character (1L), argument for queryPSICQUIC, default is "tab25"

clean
logical (1L), if TRUE extract specific information using cleanMITAB, default is TRUE

protein_only
logical (1L), if TRUE the interaction participants are restricted to proteins (exclude other types of molecules such as RNA or small molecules), default is TRUE

directory
directory where to store the data, if NULL the data is stored in <R-package-library>/MItools/data

releaseORdate
character, if data has already been downloaded: which IntAct release or download date to read

remove_obsolete_id
logical (1L), remove interactions in which one of the partners is encoded as obsolete UniProtKB accession (ID), not implemented properly: will never finish.

Value

object of class `input class`_interSpeciesInteractome containing data.table containing molecular interaction data in either of these two formats:

if clean is TRUE: contains columns as described in cleanMITAB;

if clean is FALSE: contains a standard set of columns for MITAB2.5 or MITAB2.7 depending on format;

Details

taxid1 and taxid2 is used to query specified database using PSICQUIC client, only interacting pairs between taxid1 and taxid2 are retured (no interactions within the same species, "(taxidA:taxid1 AND taxidB:taxid2) OR (taxidA:taxid2 AND taxidB:taxid1)").

interSpeciesInteractome can be used to retrive interactome data using PSICQUIC service using queryPSICQUIC, clean and select specific columns using cleanMITAB and filter resulting dataset for protein-protein interaction only. This is the default option.

Alternatively, interSpeciesInteractome can only retrive interactome data using PSICQUIC service without cleaning of filtering.

Another option is to supply MITABdata to be cleaned and filtered

Finally, you can avoid using PSICQUIC service and download data from IntAct ftp by selecting database argument "IntActFTP". This is much faster but larger requires larger download and is more computationally intensive for processing. As of 7.09.2017 "IntActFTP" provides access to DIP data, while "imex" doesn't. If database "IntActFTP" is chosen only MITAB2.7 is available and format is ignored

See also

Examples

{
# retrive a full set of human (9606) protein-protein interactions from IMEx databases in MITAB2.5 format, clean and select specific columns
interSpecies = interSpeciesInteractome(taxid1 = 9606,  taxid2 = 10239, database = "IntActFTP", format = "tab27", clean = TRUE, protein_only = TRUE)

# retrive a full set of human (9606) protein-protein interactions from IMEx databases in MITAB2.7 format not using PSICQUIC (using IntAct ftp), clean and select specific columns; save it to the specific directory inside working directory
interSpecies = interSpeciesInteractome(taxid1 = 9606,  taxid2 = 10239, database = "IntActFTP", format = "tab27", clean = TRUE, protein_only = TRUE, directory = "./data/")
}

#> ... looking for the date of the latest IntAct release ...

#> ... looking for the date of the latest IntAct release ...

#> ... loading local copy ...

#> Warning: Found and resolved improper quoting out-of-sample. First healed line 25383: <<uniprotkb:P16054	uniprotkb:Q05769	intact:EBI-298451	intact:EBI-298933|uniprotkb:Q543K3	psi-mi:kpce_mouse(display_long)|uniprotkb:Prkce(gene name)|psi-mi:Prkce(display_short)|uniprotkb:Pkce(gene name synonym)|uniprotkb:Pkcea(gene name synonym)|uniprotkb:nPKC-epsilon(gene name synonym)	psi-mi:pgh2_mouse(display_long)|uniprotkb:Ptgs2(gene name)|psi-mi:Ptgs2(display_short)|uniprotkb:Cox-2(gene name synonym)|uniprotkb:Cox2(gene name synonym)|uniprotkb:Pghs-b(gene name synonym)|uniprotkb:Tis10(gene na>>. If the fields are not quoted (e.g. field separator does not appear within any field), try quote="" to avoid this warning.

#> Warning: Found and resolved improper quoting out-of-sample. First healed line 13569: <<1857845		Spo1virus			"SPO1-like viruses"; SP01-like viruses; SPO1-like viruses; Spounalikevirus		Genus	Viruses; dsDNA viruses, no RNA stage; Caudovirales; Myoviridae; Spounavirinae	857473	>>. If the fields are not quoted (e.g. field separator does not appear within any field), try quote="" to avoid this warning.

#> Warning: Found and resolved improper quoting out-of-sample. First healed line 13569: <<1857845		Spo1virus			"SPO1-like viruses"; SP01-like viruses; SPO1-like viruses; Spounalikevirus		Genus	Viruses; dsDNA viruses, no RNA stage; Caudovirales; Myoviridae; Spounavirinae	857473	>>. If the fields are not quoted (e.g. field separator does not appear within any field), try quote="" to avoid this warning.

#> ... looking for the date of the latest IntAct release ...

#> Warning: cannot create dir './data/IntActRelease_2019Mar23', reason 'No such file or directory'

#> ... looking for the date of the latest IntAct release ...

#> ... dowloading from IntAct ftp ...

#> Warning: URL ftp://ftp.ebi.ac.uk/pub/databases/intact/current/psimitab/intact.zip: cannot open destfile './data/IntActRelease_2019Mar23/intact2019Mar23.zip', reason 'No such file or directory'

#> Warning: download had nonzero exit status

#> Warning: error 1 in extracting from zip file

#> Error in compressFile.default(filename = filename, ..., ext = ext, FUN = FUN): No such file: ./data/IntActRelease_2019Mar23/intact.txt