Split interaction subset from 2 sets so that each elements contains only one of the proteins from set 2, and retrieve FASTA sequences for each

Split interaction subset from 2 sets so that each elements contains only one of the proteins from set 2, and retrieve FASTA sequences for each

listSingleInteractFromSet2(subset1, single_interact_from_set2, set1_only,
  fasta)

Arguments

subset1
interaction_subset_from_2_sets object produced by subset2setsBy1ID

single_interact_from_set2
logical, split sequence sets to contain only one protein from interaction_set2 (only one query protein for QSLIMFinder). If FALSE, set2 will contain all proteins that interact with an element of seed_id_vect (which means multiple query proteins for QSLIMFinder).

set1_only
logical, only relevant if single_interact_from_set2 = TRUE, sequence set1 should contain only proteins that interact with an element of seed_id_vect in interaction_set1. If FALSE, proteins that interact with an element of seed_id_vect in interaction_set2 with the exception of the single one () are also included. Argument for singleInteractFromSet2

fasta
list containing AAStringSet renamed by recodeFASTA and mapping table from recodeFASTA

Value

object of class InteractionSubsetFASTA_list containing: FASTA sequences for interacting proteins, molecular interaction data they correspond to. Each element of a list contains input for individual QSLIMFinder run.

See also

Examples

subset1_fasta = listSingleInteractFromSet2(subset1, single_interact_from_set2, set1_only, fasta)

#> Error in listSingleInteractFromSet2(subset1, single_interact_from_set2,     set1_only, fasta): object 'single_interact_from_set2' not found