load Bioplex data

loadBioplex saves and/or loads the data stored on Bioplex website http://bioplex.hms.harvard.edu/downloadInteractions.php. Depends on the tsv file containing the following columns: Gene A Gene B Uniprot A Uniprot B Symbol A Symbol B p(Wrong) p(No Interaction) p(Interaction)

loadBioplex(dir = "./",
  url = "http://bioplex.hms.harvard.edu/data/BioPlex_2.3_interactionList.tsv",
  uniprot_id = T)

Arguments

dir
directory where to save/look for the local copy

uniprot_id
If TRUE uses UniprotKB accessions to identify interacting partners. If FALSE uses gene id (ENTREZ gene id) to identify interacting partners.

Value

loadBioplex saves loadBioplex data to a file, returns the object of class clean_MItab25

Examples

loadBioplex("./")

#> ... dowloading from http://bioplex.hms.harvard.edu/ ...

#> 
#> ` Object of class clean_MItab27 (molecular interaction data), for query, file, format, databases, date: `
#> [1] "Summary of protein interactions in the BioPlex network following reanalysis of all data from BioPlex 2.0 plus an additional ~1500 unpublished AP-MS experiments: http://bioplex.hms.harvard.edu/data/BioPlex_2.3_interactionList.tsv"
#> 
#> ` view of the $data: `
#>        IDs_interactor_A IDs_interactor_B interactor_IDs_databases_A
#>     1:           A5A3E0           P00813                  uniprotkb
#>     2:           Q07021           Q8N7W2                  uniprotkb
#>     3:           O75096           Q8N7W2                  uniprotkb
#>     4:           P11142           Q6ZMN8                  uniprotkb
#>     5:           Q53FA3           Q6ZMN8                  uniprotkb
#>    ---                                                             
#> 73887:           A6NHX0           Q8WTX7                  uniprotkb
#> 73888:           Q8WTX7           Q9NXC5                  uniprotkb
#> 73889:           Q6PJI9           Q8WTX7                  uniprotkb
#> 73890:           Q8WTX7         Q96S15-2                  uniprotkb
#> 73891:           Q8WTX7           Q96GY0                  uniprotkb
#>        interactor_IDs_databases_B Taxid_interactor_A Taxid_interactor_B
#>     1:                  uniprotkb               9606               9606
#>     2:                  uniprotkb               9606               9606
#>     3:                  uniprotkb               9606               9606
#>     4:                  uniprotkb               9606               9606
#>     5:                  uniprotkb               9606               9606
#>    ---                                                                 
#> 73887:                  uniprotkb               9606               9606
#> 73888:                  uniprotkb               9606               9606
#> 73889:                  uniprotkb               9606               9606
#> 73890:                  uniprotkb               9606               9606
#> 73891:                  uniprotkb               9606               9606
#>                                                                                                                     Publication_Identifiers
#>     1:        Bioplex|GeneA|100|GeneB|728378|SymbolA|ADA|SymbolB|POTEF|p(Wrong)|7.90248563407152e-10|p(No Interaction)|0.000143448545513807
#>     2:       Bioplex|GeneA|222389|GeneB|708|SymbolA|BEND7|SymbolB|C1QBP|p(Wrong)|2.40536303599679e-14|p(No Interaction)|0.00548411916162282
#>     3:      Bioplex|GeneA|222389|GeneB|4038|SymbolA|BEND7|SymbolB|LRP4|p(Wrong)|6.18626306238738e-14|p(No Interaction)|0.000167298651560179
#>     4:       Bioplex|GeneA|645121|GeneB|3312|SymbolA|CCNI2|SymbolB|HSPA8|p(Wrong)|2.66331806621062e-15|p(No Interaction)|0.0786165996748185
#>     5:       Bioplex|GeneA|645121|GeneB|3305|SymbolA|CCNI2|SymbolB|HSPA1L|p(Wrong)|1.26158784283072e-12|p(No Interaction)|0.107220558595753
#>    ---                                                                                                                                     
#> 73887: Bioplex|GeneA|652968|GeneB|729438|SymbolA|GATSL3|SymbolB|GATSL2|p(Wrong)|8.05266651996511e-10|p(No Interaction)|8.31827807081233e-10
#> 73888:    Bioplex|GeneA|652968|GeneB|54468|SymbolA|GATSL3|SymbolB|MIOS|p(Wrong)|7.25737774786179e-10|p(No Interaction)|1.34274929165835e-05
#> 73889:    Bioplex|GeneA|652968|GeneB|79726|SymbolA|GATSL3|SymbolB|WDR59|p(Wrong)|1.61686742345196e-15|p(No Interaction)|0.00661319178346156
#> 73890:   Bioplex|GeneA|652968|GeneB|84219|SymbolA|GATSL3|SymbolB|WDR24|p(Wrong)|1.85938569410255e-11|p(No Interaction)|6.95242579374724e-06
#> 73891: Bioplex|GeneA|652968|GeneB|51101|SymbolA|GATSL3|SymbolB|FAM164A|p(Wrong)|3.46729696809162e-09|p(No Interaction)|4.13482211503689e-08
#>        Confidence_values Host_organisms bait_prey_status_A bait_prey_status_B
#>     1:         0.9998566             NA                 NA                 NA
#>     2:         0.9945159             NA                 NA                 NA
#>     3:         0.9998327             NA                 NA                 NA
#>     4:         0.9213834             NA                 NA                 NA
#>     5:         0.8927794             NA                 NA                 NA
#>    ---                                                                       
#> 73887:         1.0000000             NA                 NA                 NA
#> 73888:         0.9999866             NA                 NA                 NA
#> 73889:         0.9933868             NA                 NA                 NA
#> 73890:         0.9999930             NA                 NA                 NA
#> 73891:         1.0000000             NA                 NA                 NA
#>        Interaction_detection_methods Interaction_types Interaction_identifiers
#>     1:                            NA                NA                      NA
#>     2:                            NA                NA                      NA
#>     3:                            NA                NA                      NA
#>     4:                            NA                NA                      NA
#>     5:                            NA                NA                      NA
#>    ---                                                                        
#> 73887:                            NA                NA                      NA
#> 73888:                            NA                NA                      NA
#> 73889:                            NA                NA                      NA
#> 73890:                            NA                NA                      NA
#> 73891:                            NA                NA                      NA
#>        Expansion_methods Features_interactor_A Features_interactor_B
#>     1:                NA                    NA                    NA
#>     2:                NA                    NA                    NA
#>     3:                NA                    NA                    NA
#>     4:                NA                    NA                    NA
#>     5:                NA                    NA                    NA
#>    ---                                                              
#> 73887:                NA                    NA                    NA
#> 73888:                NA                    NA                    NA
#> 73889:                NA                    NA                    NA
#> 73890:                NA                    NA                    NA
#> 73891:                NA                    NA                    NA
#>        Identification_method_participant_A Identification_method_participant_B
#>     1:                                  NA                                  NA
#>     2:                                  NA                                  NA
#>     3:                                  NA                                  NA
#>     4:                                  NA                                  NA
#>     5:                                  NA                                  NA
#>    ---                                                                        
#> 73887:                                  NA                                  NA
#> 73888:                                  NA                                  NA
#> 73889:                                  NA                                  NA
#> 73890:                                  NA                                  NA
#> 73891:                                  NA                                  NA
#>        binding_region_A_start binding_region_A_end binding_region_B_start
#>     1:                     NA                   NA                     NA
#>     2:                     NA                   NA                     NA
#>     3:                     NA                   NA                     NA
#>     4:                     NA                   NA                     NA
#>     5:                     NA                   NA                     NA
#>    ---                                                                   
#> 73887:                     NA                   NA                     NA
#> 73888:                     NA                   NA                     NA
#> 73889:                     NA                   NA                     NA
#> 73890:                     NA                   NA                     NA
#> 73891:                     NA                   NA                     NA
#>        binding_region_B_end binding_region_A_type binding_region_B_type
#>     1:                   NA                    NA                    NA
#>     2:                   NA                    NA                    NA
#>     3:                   NA                    NA                    NA
#>     4:                   NA                    NA                    NA
#>     5:                   NA                    NA                    NA
#>    ---                                                                 
#> 73887:                   NA                    NA                    NA
#> 73888:                   NA                    NA                    NA
#> 73889:                   NA                    NA                    NA
#> 73890:                   NA                    NA                    NA
#> 73891:                   NA                    NA                    NA
#>                pair_id
#>     1:   A5A3E0|P00813
#>     2:   Q07021|Q8N7W2
#>     3:   O75096|Q8N7W2
#>     4:   P11142|Q6ZMN8
#>     5:   Q53FA3|Q6ZMN8
#>    ---                
#> 73887:   A6NHX0|Q8WTX7
#> 73888:   Q8WTX7|Q9NXC5
#> 73889:   Q6PJI9|Q8WTX7
#> 73890: Q8WTX7|Q96S15-2
#> 73891:   Q8WTX7|Q96GY0