Use PSICQUIC to query for molecular interactions (using MIQL, specifying database and output format)

query_PSICQUIC function queries a specified molecular interaction database using MIQL query language, retrieves the data in a specified format and saves it to a file

queryPSICQUIC(query, format, database, file)

Arguments

query
search query in MIQL query language, as you would type in PSICQUIC View client: http://www.ebi.ac.uk/Tools/webservices/psicquic/view/home.xhtml

format
output format (most widely used tabular format is tab25, use tab27 for more data columns)

database
PSICQUIC service (full list: http://www.ebi.ac.uk/Tools/webservices/psicquic/registry/registry?action=STATUS), use "imex" shorthand to choose all IMEx databases

file
file name and directory to save the result ("/dir/filename")

Value

saves query result to a file, returns the query settings, how many interactions per database retrieved and which databases are inactive

Details

See https://psicquic.github.io/MiqlReference.html or https://psicquic.github.io/MiqlReference27.html for the description of MIQL query language.

Unlike rawQuery function from PSICQUIC package this function allows to use all IMEx databases by passing "imex" to database argument and also separates the choice of output format from the query. Output format options: https://psicquic.github.io/formats.html

This function also splits large query result into chunks (2500 interactions) to limit the load on the PSICQUIC servers.

List of data provider names accepted by the function: "imex", "APID Interactomes", "BioGrid", "bhf-ucl", "ChEMBL", "HPIDb", "InnateDB", "InnateDB-All", "IntAct", "mentha", "MPIDB", "MatrixDB", "MINT", "Reactome", "Reactome-FIs", "EBI-GOA-miRNA", "I2D", "I2D-IMEx", "InnateDB-IMEx", "MolCon", "UniProt", "MBInfo", "VirHostNet", "BAR", "EBI-GOA-nonIntAct", "ZINC"

Examples

queryPSICQUIC(query = "id:P74565 AND detmethod:\"MI:0018\"",
               format = "tab27",
               database = "imex",
               file = "P74565_2H_interactions_imex_tab27.tsv")

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: number of items to replace is not a multiple of replacement length

#>                                     query                                  file
#>  1: id:P74565 AND detmethod:%22MI:0018%22 P74565_2H_interactions_imex_tab27.tsv
#>  2: id:P74565 AND detmethod:%22MI:0018%22 P74565_2H_interactions_imex_tab27.tsv
#>  3: id:P74565 AND detmethod:%22MI:0018%22 P74565_2H_interactions_imex_tab27.tsv
#>  4: id:P74565 AND detmethod:%22MI:0018%22 P74565_2H_interactions_imex_tab27.tsv
#>  5: id:P74565 AND detmethod:%22MI:0018%22 P74565_2H_interactions_imex_tab27.tsv
#>  6: id:P74565 AND detmethod:%22MI:0018%22 P74565_2H_interactions_imex_tab27.tsv
#>  7: id:P74565 AND detmethod:%22MI:0018%22 P74565_2H_interactions_imex_tab27.tsv
#>  8: id:P74565 AND detmethod:%22MI:0018%22 P74565_2H_interactions_imex_tab27.tsv
#>  9: id:P74565 AND detmethod:%22MI:0018%22 P74565_2H_interactions_imex_tab27.tsv
#> 10: id:P74565 AND detmethod:%22MI:0018%22 P74565_2H_interactions_imex_tab27.tsv
#> 11: id:P74565 AND detmethod:%22MI:0018%22 P74565_2H_interactions_imex_tab27.tsv
#>     format all.databases n.interactions.in.database database.not.active
#>  1:  tab27        IntAct                         44                    
#>  2:  tab27          MINT                          0                    
#>  3:  tab27       bhf-ucl                          0                    
#>  4:  tab27         MPIDB                          0                    
#>  5:  tab27      MatrixDB                     <html>                    
#>  6:  tab27         HPIDb                          0                    
#>  7:  tab27      I2D-IMEx                                       I2D-IMEx
#>  8:  tab27 InnateDB-IMEx                                  InnateDB-IMEx
#>  9:  tab27        MolCon                                         MolCon
#> 10:  tab27       UniProt                          0                    
#> 11:  tab27        MBInfo                          0                    

queryPSICQUIC(query = "id:P74565 AND detmethod:\"MI:0018\"",
               format = "tab25",
               database = "mentha",
               file = "P74565_2H_interactions_mentha_tab25.tsv")

#>                                    query
#> 1: id:P74565 AND detmethod:%22MI:0018%22
#>                                       file format all.databases
#> 1: P74565_2H_interactions_mentha_tab25.tsv  tab25        mentha
#>    n.interactions.in.database database.not.active
#> 1:                         44                    

queryPSICQUIC(query = "id:P74565",
               format = "tab25",
               database = "mentha",
               file = "P74565_2H_interactions_mentha_tab25.tsv")

#>        query                                    file format all.databases
#> 1: id:P74565 P74565_2H_interactions_mentha_tab25.tsv  tab25        mentha
#>    n.interactions.in.database database.not.active
#> 1:                         44                    

queryPSICQUIC(query = "id:156",
               format = "tab25",
               database = "BioGrid",
               file = "entrezgene156_interactions_BioGrid_tab25.tsv")

#>     query                                         file format all.databases
#> 1: id:156 entrezgene156_interactions_BioGrid_tab25.tsv  tab25       BioGrid
#>    n.interactions.in.database database.not.active
#> 1:                         82