Run CompariMotif3

Run CompariMotif3

runCompariMotif3(input_file = "./SLIMFinder/result/motifs.txt",
  slimpath = "../software/slimsuite/tools/", dbpath = "./data_files/",
  dburl = "http://elm.eu.org/elms/elms_index.tsv",
  parameters = "unmatched=T motific=T",
  out_file = "./SLIMFinder/result/comparimotif.tdt",
  LSF_project_path = "/hps/nobackup/research/petsalaki/users/vitalii/vitalii/viral_project/",
  run = F, with = c("self", "db")[2])

Arguments

input_file
a path to a tsv file containing Name and Pattern columns

slimpath
path to /slimsuite/tools/

dbpath
path to directory where to save and keep ELM database (http://elm.eu.org/) or other database of linear motifs in a format required by comparimotif_V3: http://rest.slimsuite.unsw.edu.au/docs&page=module:comparimotif_V3

dburl
url where to download database

parameters
any commandline parameters for comparimotif_V3 as a single character string

out_file
path to a file where comparimotif_V3 should save the result

LSF_project_path
path to the project directory on the LSF cluster file system. If on LSF is to be used all other paths should be relative to this directory, alternatively, set this argument to empty character. Final path is LSF_project_path concatenated with input_file, slimpath, out_file or dbpath. When using this on a local machine set this argument to empty character.

run
logical, run comparimotif_V3 (T) or just output the command (F), can be useful for check correcteness

with
compare motifs in input_file to the database in dbpath ("db") or to input_file ("self")

Value

system command to be run or that was run

See also