Run a list of QSLIMFinder jobs

Run a list of QSLIMFinder jobs

runQSLIMFinder(commands_list, file_list, max_memory = 5000, onLSF = T,
  recursive = F, lsf_keyword = "TERM_MEMLIMIT", rm_log = T,
  memory_step = 100, memory_start = 200, Njobs_limit = 490)

modifyMemoryInBsub(commands, memory)

LSFrunQSLIMFinder(commands, job_name_sig = "batch_", Njobs_limit = 490)

jobsCrashed(commands_list, rm_log = T, lsf_keyword = "TERM_MEMLIMIT")

Arguments

commands_list
list containing: 1. command to set up enviromental variable IUPred_PATH; 2. character vector of bash commands that will lauch QSLIMFinder as a job on LSF cluster; 3, 4, 5 - directories where LSF should write stout and sterr

file_list
data.table containing path to files and directories for QSLIMFinder: fastafile, queryfile, outputdir, outputfile

max_memory
max memory to use for a job

onLSF
use LSF cluster to run QSLIMFinder jobs

recursive
how to create directories

lsf_keyword
character, LSF termination reason keyword (https://www.ibm.com/support/knowledgecenter/en/SSETD4_9.1.3/lsf_admin/termination_reasons_lsf.html). Defaults to "TERM_MEMLIMIT"

rm_log
remove log after checking for termination reason

Njobs_limit
integer, the number of LSF jobs allowed to run simultaneously

commands
character vector of commands generated by QSLIMFinderCommand that will launch QSLIMFinder jobs, commands_list[[2]]

memory
integer, how much memory should the job use

job_name_sig
string common to all job names

Value

list: character vector, commands that have crashed even at max_memory; data.table, corresponding rows of file_list

jobsCrashed(): vector of commands that crashed

See also