Retrieve molecular interactions for the random set of proteins (of a particular taxon)

Retrieve molecular interactions for the random set of proteins (of a particular taxon)

randomInteractome(MITABdata = NULL, degree_data = NULL, n_prot,
  degree_dist = NULL, ...)

Arguments

degree_data
data.table containing pre-calculated (using edgelist2degree) degree for each node in MITABdata, usefull for taking multiple samples, the default in NULL

n_prot
integer (1L), the number of proteins for which to retrieve the random set of interactions

degree_dist
data.table, specifies the degree frequency () for each degree (N) to produce the network with the specific degree distribution, if set to NULL (default) the degree distribution will correspond to that of taxid interactome

...
Arguments passed on to fullInteractome

MITABdata

object of class "RAW_MItab25" or "RAW_MItab27" (list) containing molecular interaction data as returned by queryPSICQUICrlib or loadIntActFTP, default in NULL

taxid

character (1L), taxonomy id of the species which interaction participants should belong to, default is "9606" (which is human)

database

character (1L), argument for queryPSICQUIC, PSICQUIC-compliant database to query for interactions. The default is "imex" alternative to which is "IntActFTP"

protein_only

logical (1L), if TRUE the interaction participants are restricted to proteins (exclude other types of molecules such as RNA or small molecules), default is TRUE

directory

directory where to store the data, if NULL the data is stored in <R-package-library>/MItools/data

releaseORdate

character, if data has already been downloaded: which IntAct release or download date to read

remove_obsolete_id

logical (1L), remove interactions in which one of the partners is encoded as obsolete UniProtKB accession (ID), not implemented properly: will never finish.

within_species

logical (1L), return interactions only between proteins of taxid. If FALSE returns interactions for taxid proteins both within taxid and with proteins from other species

Value

list of two elements: 1. interactome - data.table containing molecular interaction data in either of these two formats:

if clean is TRUE: contains columns as described in cleanMITAB;

if clean is FALSE: contains a standard set of columns for MITAB2.5 or MITAB2.7 depending on format;

2. seed - character vector containing IDs of proteins used as a seed to retrieve molecular interactions

Details

Not implemented: Random network can be specified to have specific degree distribution. If the (degree parameter is set taxid proteins will be split by degree and from each degree group a sample of the size specified by how many times specific degree number is repeated in degree will be taken.

If the degree distribution is not specified a sample of n_prot is taken from all proteins which have interaction data available in the database for taxid. In this case, the degree distribution of the resulting set of proteins will be similar to the degree distribution in the interactome of taxid in database.

randomInteractome retrieves molecular interactions using fullInteractome

Examples

# retrive the interactome using PSICQIUC servise (or by reading local copy) from IMEx databases for a list of 200 random human (9606) proteins, not specifying their degree distribution
set.seed(1)
random = randomInteractome(n_prot = 200, degree_dist = NULL, taxid = "9606", database = "imex", protein_only = TRUE)

#> ... looking for the date of the latest IntAct release ...

#> found local copy of the data from the latest release ... reading into R

# retrive the interactome from MITABdata for a list of 200 random human (9606) proteins, not specifying their degree_dist distribution
full = fullInteractome(taxid = "9606", database = "imex", format = "tab25", clean = TRUE, protein_only = TRUE)

#> ... looking for the date of the latest IntAct release ...

#> found local copy of the data from the latest release ... reading into R

set.seed(1)
random = randomInteractome(MITABdata = full, n_prot = 200, degree_dist = NULL)

# retrive the interactome using PSICQIUC servise (or by reading local copy from a specified directory) from IMEx databases for a list of 200 random human (9606) proteins, not specifying their degree_dist distribution
set.seed(1)
random = randomInteractome(n_prot = 200, degree_dist = NULL, taxid = "9606", database = "imex", protein_only = TRUE, directory = "./data/")

#> ... looking for the date of the latest IntAct release ...

#> Warning: cannot create dir './data/IntActRelease_2018Aug7', reason 'No such file or directory'

#> Error in queryPSICQUICrlib(query = query, format = format, database = database,     directory = directory, releaseORdate = releaseORdate): dir.exists(pkg_dir_last_release) is not TRUE